get owner data not working

+++ app/src/main/java/uz/gita/dtm/data/models/mapper/Mapper.kt
@@ -47,7 +47,7 @@ object Mapper {
         image = this["image"].toString(),
         number = this["number"].toString(),
         date = this["date"].toString().toLong(),
-        time = this["time"].toString().toLong(),
+        time = this["time"].toString(),
         state = this["state"].toString().toBoolean(),
         pay = this["pay"].toString(),
         applicantName = this["applicant_name"].toString(),
